Fixed styling on create account page
