+++ frontend-Idata2301/code/app/src/App.jsx
@@ -6,9 +6,9 @@ import ProductCardHeader from "./components/cards/deals/ProductCardHeader";
 import FlightsContainer from "./components/cards/searched-flights/FlightsContainer";
 import MainPageHero from "./components/hero/mainPageHero/MainPageHero";
 import DealsPageHero from "./components/hero/dealsPageHero/DealsPageHero";
-import LogInPageHero from "./components/hero/logInPageHero/LogInPageHero";
+import LogInPageHero from "./components/hero/account/logInPageHero/LogInPageHero";
 import ShoppingCartHero from "./components/hero/shoppingCartPageHero/shoppingCartHero";
-import CreateAccount from "./pages/createAccount";
+import CreateAccount from "./components/hero/account/createAccountHero/createAccountHero";
 import Footer from "./components/footer/Footer";
 import Navbar from "./components/header/Navbar";
 import AboutUs from "./components/about/AboutUs";
