+++ frontend-Idata2301/code/app/src/components/hero/account/createAccountHero/createAccountHero.html
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Create Account</title>
+  <link rel="stylesheet" href="createAccountPage.css" />
+</head>
+<body>
+  <section class="hero">
+    <div class="hero-container">
+      <h1>Welcome to Flight Finder!</h1>
+      <p>Please create an account:</p>
+
+      <form action="/create-account" method="post">
+        <div class="form-group">
+          <label for="username">Username</label>
+          <input type="text" id="username" name="username" required />
+        </div>
+
+        <div class="form-group">
+          <label for="email">E-post</label>
+          <input type="email" id="email" name="email" required />
+        </div>
+
+        <div class="form-group">
+          <label for="password">Passord</label>
+          <input type="password" id="password" name="password" required />
+        </div>
+
+        <div class="form-group">
+          <label for="confirm-password">Bekreft passord</label>
+          <input type="password" id="confirm-password" name="confirm-password" required />
+        </div>
+
+        <button type="submit" class="primary-button">Opprett konto</button>
+      </form>
+    </div>
+  </section>
+</body>
+</html>
