--- frontend-Idata2301/code/app/src/components/hero/createAccountHero/createAccountHero.css
@@ -1,40 +1,0 @@
-.hero.login-hero-container {
-    background-size: cover;
-    padding: 50px 0;
-    text-align: center;
-    color: white;
-  }
-  .login-hero-container {
-    max-width: 600px;
-    margin: 0 auto;
-    background: rgba(0, 0, 0, 0.5);
-    padding: 20px;
-    border-radius: 10px;
-    z-index: 2;
-  }
-  .form-group.login-hero-container {
-    margin-bottom: 15px;
-  }
-  label.login-hero-container {
-    display: block;
-    margin-bottom: 5px;
-  }
-  input.login-hero-container {
-    width: 100%;
-    padding: 10px;
-    border: none;
-    border-radius: 5px;
-    z-index: 3;
-  }
-  button.login-hero-container {
-    padding: 10px 20px;
-    border: none;
-    border-radius: 5px;
-    background: #007BFF;
-    color: white;
-    cursor: pointer;
-    z-index: 3;
-  }
-  button:hover {
-    background: #0056b3;
-  }
