--- frontend-Idata2301/code/app/src/components/hero/createAccountHero/createAccountHero.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<section class="hero">
-    <div class="hero-container">
-      <h1>Welcome To Flight Finder!</h1>
-      <p>Please create an account!</p>
-      <form action="/create-account" method="post">
-        <div class="form-group">
-          <label for="username">Username:</label>
-          <input type="text" id="username" name="username" required>
-        </div>
-        <div class="form-group">
-            <label for="email">Email:</label>
-            <input type="email" id="email" name="email" required>
-        </div>
-        <div class="form-group">
-            <label for="password">Password:</label>
-            <input type="password" id="password" name="password" required>
-        </div>
-        <div class="form-group">
-            <label for="confirm-password">Confirm Password:</label>
-            <input type="password" id="confirm-password" name="confirm-password" required>
-        </div>
-        <button type="submit">Create Account</button>
-      </form>
-    </div>
-  </section>
